Move maps to postgres using Sequel
Rather than deal directly with writing SQL, we use the Sequel library.
We replace all mongo calls in the map endpoints with this, and clean up
the map scraper - renaming lng, as well as removing the dead image_url
code. In addition, it cleans up the typing of the response for v1
endpoints.

Resolves #120. Supersedes #121.