Convert bus routes to Sequel; Kill mongodb (#141)
This converts the bus routes over to use Sequel, greatly simplifying a lot of code, and letting us finally delete MongoDB. Ideally, no route responses have changed, and this will be double checked again.

The only flaw in this is that the database is not actually relational, and all we've really done is replicate the same behavior under a different framework. However, this at least gives us a new base to work off of.